removed deprecated NewElementXD methods, and moved NewInternalElement2D methods to Query class and folders
